+++ 2017code/Tests/org/usfirst/frc/team3328/robot/SteamWorksDriveSystemTest.java
@@ -2,16 +2,19 @@ package org.usfirst.frc.team3328.robot;
 
 import static org.junit.Assert.*;
 import org.junit.Test;
+import org.usfirst.frc.team3328.robot.subsystems.SteamWorksDriveSystem;
+import org.usfirst.frc.team3328.robot.utilities.DriveEncoders;
 
 public class SteamWorksDriveSystemTest {
 	
+	DriveEncoders encoders = new DriveEncoders();
 	FakeController fakeCont = new FakeController();
 	FakeSpeedController fl = new FakeSpeedController();
 	FakeSpeedController fr = new FakeSpeedController();
 	FakeSpeedController bl = new FakeSpeedController();
 	FakeSpeedController br = new FakeSpeedController();
 	FakeADIS16448_IMU imu = new FakeADIS16448_IMU();
-	SteamWorksDriveSystem drive = new SteamWorksDriveSystem(fl, fr, bl, br, fakeCont);
+	SteamWorksDriveSystem drive = new SteamWorksDriveSystem(encoders, fl, fr, bl, br, fakeCont);
 
 	@Test
 	public void controlledMove_yLargerThanX_rightMotorTurnsBackwards() {
