+++ 2017code/src/org/usfirst/frc/team3328/robot/networking/Target.java
@@ -1,4 +1,4 @@
-package org.usfirst.frc.team3328.robot;
+package org.usfirst.frc.team3328.robot.networking;
 
 import java.io.Serializable;
